+++ tools/tpm2_ecephemeral.c
@@ -7,6 +7,7 @@
 #include "tpm2_alg_util.h"
 #include "tpm2_options.h"
 
+#define MAX_SESSIONS 3
 typedef struct tpm_ecephemeral_ctx tpm_ecephemeral_ctx;
 struct tpm_ecephemeral_ctx {
     /*
@@ -21,15 +22,31 @@ struct tpm_ecephemeral_ctx {
     TPM2B_ECC_POINT *Q;
     char *commit_counter_path;
     char *ephemeral_pub_key_path;
+
+    /*
+     * Parameter hashes
+     */
+    const char *cp_hash_path;
+    TPM2B_DIGEST cp_hash;
+    bool is_command_dispatch;
+    TPMI_ALG_HASH parameter_hash_algorithm;
+
 };
 
 static tpm_ecephemeral_ctx ctx = {
     .curve_id = TPM2_ECC_NONE,
+    .parameter_hash_algorithm = TPM2_ALG_ERROR,
 };
 
 static tool_rc ecephemeral(ESYS_CONTEXT *ectx) {
 
-    return tpm2_ecephemeral(ectx, ctx.curve_id, &ctx.Q, &ctx.counter);
+    tool_rc rc = tpm2_ecephemeral(ectx, ctx.curve_id, &ctx.Q, &ctx.counter,
+        &ctx.cp_hash, ctx.parameter_hash_algorithm);
+    if (rc != tool_rc_success) {
+        LOG_ERR("Failed TPM2_CC_EC_Ephemeral");
+    }
+
+    return rc;
 }
 
 static tool_rc process_outputs(ESYS_CONTEXT *ectx) {
@@ -39,12 +56,25 @@ static tool_rc process_outputs(ESYS_CONTEXT *ectx) {
     /*
      * 1. Outputs that do not require TPM2_CC_<command> dispatch
      */
+    bool is_file_op_success = true;
+    if (ctx.cp_hash_path) {
+        is_file_op_success = files_save_digest(&ctx.cp_hash, ctx.cp_hash_path);
+
+        if (!is_file_op_success) {
+            return tool_rc_general_error;
+        }
+    }
+
+    tool_rc rc = tool_rc_success;
+    if (!ctx.is_command_dispatch) {
+        return rc;
+    }
 
     /*
      * 2. Outputs generated after TPM2_CC_<command> dispatch
      */
     FILE *fp = fopen(ctx.commit_counter_path, "wb");
-    bool is_file_op_success = files_write_16(fp, ctx.counter);
+    is_file_op_success = files_write_16(fp, ctx.counter);
     fclose(fp);
     if (!is_file_op_success) {
         LOG_ERR("Failed to write out the ECC commit count");
@@ -90,10 +120,20 @@ static tool_rc process_inputs(ESYS_CONTEXT *ectx) {
     /*
      * 4.a Determine pHash length and alg
      */
+    tpm2_session *all_sessions[MAX_SESSIONS] = {
+        0,
+        0,
+        0
+    };
 
+        const char **cphash_path = ctx.cp_hash_path ? &ctx.cp_hash_path : 0;
+
+    ctx.parameter_hash_algorithm = tpm2_util_calculate_phash_algorithm(ectx,
+        cphash_path, &ctx.cp_hash, 0, 0, all_sessions);
     /*
      * 4.b Determine if TPM2_CC_<command> is to be dispatched
      */
+    ctx.is_command_dispatch = ctx.cp_hash_path ? false : true;
 
     return tool_rc_success;
 }
@@ -127,6 +167,9 @@ static bool on_option(char key, char *value) {
     case 't':
         ctx.commit_counter_path = value;
         break;
+    case 0:
+        ctx.cp_hash_path = value;
+    break;
     };
 
     return true;
@@ -164,6 +207,7 @@ static bool tpm2_tool_onstart(tpm2_options **opts) {
     static struct option topts[] = {
       { "public",   required_argument, 0, 'u' },
       { "counter",  required_argument, 0, 't' },
+      { "cphash",   required_argument, 0,  0  },
     };
 
     *opts = tpm2_options_new("u:t:", ARRAY_LEN(topts), topts,