+++ docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### next
 
+  * tpm2_ecephemeral:
+      - Added option to output the cpHash for the command TPM2_CC_EC_Ephemeral.
+
   * tpm2_commit:
       - Added option to output the cpHash for the command TPM2_CC_Commit.